+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Add finalizers to temp file classes to catch files not deleted at application exit
 * Remove Exists() method from DicomDataset (duplicates functionality of Contains())
 * Extension methods for recalculating and removing group length elements
+* Force calculation of group lengths when writing File Meta Info and Command datasets
 
 #### v1.0.23 (09/26/2012)
 * Fix W/L calculation creating negative window width